+++ src/rst-rt/dynamics/Wrench.hpp
@@ -39,12 +39,12 @@ namespace dynamics {
 class Wrench {
 public:
 	Wrench();
-	Wrench(double x, double y, double z, double a, double b, double c);
+	Wrench(float x, float y, float z, float a, float b, float c);
 	Wrench(rstrt::dynamics::Forces f, rstrt::dynamics::Torques t);
 
 //private:
-    Eigen::Matrix<double, 3, 1> forces;
-    Eigen::Matrix<double, 3, 1> torques;
+    Eigen::Matrix<float, 3, 1> forces;
+    Eigen::Matrix<float, 3, 1> torques;
 };
 
 // Displaying:
